From review branch _reviews/master/scar/AddGuard/2 into master

+++ .delivery/build-cookbook/recipes/deploy.rb
@@ -54,7 +54,9 @@ if ['delivered'].include?(node['delivery']['change']['stage'])
       version = data_hash['Parameters']['Version']['Default']
       s3 = Aws::S3::Resource.new(region:'us-west-2')
       obj = s3.bucket('bjcpublic').object("cloudformation/bjc-demo-#{version}.json")
-      obj.upload_file("#{workspace}/stacks/acceptance-bjc-demo.json", acl:'public-read')
+      unless obj.exists?
+        obj.upload_file("#{workspace}/stacks/acceptance-bjc-demo.json", acl:'public-read')
+      end
     end
   end
 end
